requirements.txt
pillow
tqdm
art
